Fix an issue where helpers wouldn't be loaded occasionally.
